requirements.txt
openai
langchain
torch
ctransformers
transformers
langchain-community
sentence-transformers
datasets
faiss-cpu
tiktoken 
chromadb 
pypdf 
InstructorEmbedding
accelerate 
bitsandbytes 
sentencepiece 
Xformers
ragas
python-dotenv
langchain_experimental
langchain_cohere
langchain_pinecone
spacy
nltk
llama_index